# static/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Teamed Backend - Flask 3.0.2",
    "description": "Main documentation for Teamed API. Any further questions, you can click the contact link below.",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "n10808329@qut.edu.au"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.11"
  },
  "servers": [
    {
      "url": "https://teamedbackend.pythonanywhere.com/"
    }
  ],
  "tags": [
    {
      "name": "user",
      "description": "Describes different users on Teamed"
    },
    {
      "name": "lead",
      "description": "A lead or pre-worked project"
    },
    {
      "name": "project",
      "description": "Worked on prospect"
    }
  ],
  "paths": {
    "/login": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logs user into their Teamed dashboard",
        "description": "Uses email and password of user to login to dashboard.",
        "operationId": "loginUser",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "description": "The user name for login",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "password",
            "in": "query",
            "description": "The password for login in clear text",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "headers": {
              "X-Rate-Limit": {
                "description": "calls per hour allowed by the user",
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              },
              "X-Expires-After": {
                "description": "date in UTC when token expires",
                "schema": {
                  "type": "string",
                  "format": "date-time"
                }
              }
            },
            "content": {
              "application/xml": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username/password supplied"
          }
        }
      }
    },
    "/register": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Logs out current logged in user session",
        "description": "",
        "operationId": "logoutUser",
        "parameters": [],
        "responses": {
          "default": {
            "description": "successful operation"
          }
        }
      }
    },
    "/users/{email}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get user by user name",
        "description": "",
        "operationId": "getUserByName",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "The name that needs to be fetched. Use user1 for testing. ",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get all users",
        "description": "",
        "operationId": "getAllUsers",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid username supplied"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/lead/add": {
      "post": {
        "tags": [
          "lead"
        ],
        "summary": "Adds a lead to the leads tabls",
        "operationId": "addLead",
        "parameters": [
          {
            "in": "body",
            "name": "lead",
            "description": "Body request for creating a lead",
            "schema": {
              "$ref": "#/components/schemas/Lead"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully adds a lead to the Leads table",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Lead"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 10
          },
          "name": {
            "type": "string",
            "example": "theUser"
          },
          "email": {
            "type": "string",
            "example": "john@email.com"
          },
          "password": {
            "type": "string",
            "example": "12345"
          },
          "phone": {
            "type": "string",
            "example": "12345"
          },
          "userType": {
            "type": "string",
            "description": "Type of user on Teamed",
            "example": "PROJECT_MANAGER, CLIENT, FREELANCER"
          }
        },
        "xml": {
          "name": "user"
        }
      },
      "Lead": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "theUser"
          },
          "description": {
            "type": "string",
            "example": "Electrical Company"
          },
          "upper_estimate": {
            "type": "integer",
            "format": "int64",
            "example": 100000
          },
          "lower_estimate": {
            "type": "integer",
            "format": "int64",
            "example": 40000
          },
          "closing_date": {
            "type": "string",
            "description": "Type of user on Teamed",
            "example": "PROJECT_MANAGER"
          },
          "status": {
            "type": "string",
            "description": "Lead can be a prospect, submitted, successful, unsuccessful",
            "example": "prospects"
          }
        }
      }
    },
    "requestBodies": {
      "UserArray": {
        "description": "List of user object",
        "content": {
          "application/json": {
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "petstore_auth": {
        "type": "oauth2",
        "flows": {
          "implicit": {
            "authorizationUrl": "https://petstore3.swagger.io/oauth/authorize",
            "scopes": {
              "write:pets": "modify pets in your account",
              "read:pets": "read your pets"
            }
          }
        }
      },
      "api_key": {
        "type": "apiKey",
        "name": "api_key",
        "in": "header"
      }
    }
  }
}